src/requirements.txt
# packages that depend on the development environment.
setuptools
build
flake8
black
Sphinx
behave
gunicorn
flask
flask-restx
gevent
flask-socketio
flask-cors
pyOpenSSL
gevent-websocket
websocket-client
python-dotenv
colorlog
